requirements.txt
aiosignal==1.2.0
anyio==3.6.1
APScheduler==3.9.1
attrs==21.4.0
backports.zoneinfo==0.2.1
black==22.6.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
chess==1.9.2
click==8.0.4
cycler==0.11.0
distlib==0.3.4
filelock==3.7.1
fonttools==4.33.3
frozenlist==1.3.0
grpcio==1.43.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
imageio==2.19.3
imgaug==0.4.0
importlib-resources==5.8.0
imutils==0.5.4
joblib==1.1.0
jsonschema==4.6.2
kiwisolver==1.4.2
matplotlib==3.5.2
msgpack==1.0.4
mypy-extensions==0.4.3
networkx==2.8.4
numpy==1.22.4
opencv-python==4.5.5.64
packaging==21.3
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
protobuf==3.20.1
pygad==2.17.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-telegram-bot==20.0a2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==6.0
ray==1.13.0
requests==2.28.1
result==0.8.0
rfc3986==1.5.0
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.8.1
Shapely==1.8.2
six==1.16.0
sniffio==1.2.0
threadpoolctl==3.1.0
tifffile==2022.5.4
tomli==2.0.1
tornado==6.2
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
virtualenv==20.15.1
zipp==3.8.0
imagehash==4.1.0
